extract bezier control behavior to mixin
