+++ src/canvastools/ts/CanvasTools/Interface/IAnchorMixin.ts
@@ -1,0 +1,4 @@
+export interface IAnchorMixin {
+    buildAnchors(): void;
+    redraw(): void;
+}
